requirements.txt
fastapi
uvicorn
python-dotenv
openai
azure-search-documents
azure-core